# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A read-only Aave V3 data monitoring service",
        "title": "aavepulse",
        "contact": {},
        "version": "v0.1.0"
    },
    "paths": {
...
        "/api/v1/defi/overview": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "DeFi"
                ],
                "summary": "Get Aave protocol overview",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.Overview"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/defi/reserves": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "DeFi"
                ],
                "summary": "Query Aave reserves",
                "parameters": [
                    {
                        "type": "string",
                        "description": "asset symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort field",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc or desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.Reserve"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/defi/reserves/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "DeFi"
                ],
                "summary": "Get an Aave reserve",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reserve ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.Reserve"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/defi/reserves/{id}/snapshots": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "DeFi"
                ],
                "summary": "Query reserve snapshots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reserve ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "30d",
                        "description": "7d, 30d, or 90d",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.ReserveSnapshot"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/defi/sync-runs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "DeFi"
                ],
                "summary": "Query Aave synchronization runs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "DeFi"
                ],
                "summary": "Run Aave synchronization",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "schema.Overview": {
            "type": "object",
            "properties": {
                "available_liquidity_usd": {
                    "type": "number"
                },
                "demo": {
                    "type": "boolean"
                },
                "last_synced_at": {
                    "type": "string"
                },
                "reserve_count": {
                    "type": "integer"
                },
                "total_borrowed_usd": {
                    "type": "number"
                },
                "total_supplied_usd": {
                    "type": "number"
                },
                "utilization_rate": {
                    "type": "number"
                }
            }
        },
        "schema.Reserve": {
            "type": "object",
            "properties": {
                "available_liquidity_usd": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "data_updated_at": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "demo": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "market_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_usd": {
                    "type": "number"
                },
                "protocol_id": {
                    "type": "string"
                },
                "supply_apy": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "total_borrowed_usd": {
                    "type": "number"
                },
                "total_supplied_usd": {
                    "type": "number"
                },
                "underlying_asset": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "utilization_rate": {
                    "type": "number"
                },
                "variable_borrow_apy": {
                    "type": "number"
                }
            }
        },
        "schema.ReserveSnapshot": {
            "type": "object",
            "properties": {
                "available_liquidity_usd": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "demo": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "price_usd": {
                    "type": "number"
                },
                "reserve_id": {
                    "type": "string"
                },
                "snapshot_at": {
                    "type": "string"
                },
                "supply_apy": {
                    "type": "number"
                },
                "total_borrowed_usd": {
                    "type": "number"
                },
                "total_supplied_usd": {
                    "type": "number"
                },
                "utilization_rate": {
                    "type": "number"
                },
                "variable_borrow_apy": {
                    "type": "number"
                }
            }
        },
...
        "util.ResponseResult": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errors.Error"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}